add sorting + height matching and moved to utils
