+++ pkg/rpc/core/utils.go
@@ -3,12 +3,17 @@ package core
 import (
 	"context"
 	"encoding/hex"
+	"encoding/json"
 	"errors"
 	"fmt"
+	"sort"
 
 	cmttypes "github.com/cometbft/cometbft/types"
+	ds "github.com/ipfs/go-datastore"
+	dsq "github.com/ipfs/go-datastore/query"
 
 	"github.com/rollkit/go-execution-abci/pkg/common"
+	rlktypes "github.com/rollkit/rollkit/types"
 )
 
 const NodeIDByteLength = 20
@@ -103,4 +108,102 @@ func TruncateNodeID(idStr string) (string, error) {
 		return "", fmt.Errorf("node ID too short, expected at least %d bytes, got %d", NodeIDByteLength, len(idBytes))
 	}
 	return hex.EncodeToString(idBytes[:NodeIDByteLength]), nil
+}
+
+func getHeightFromEntry(field string, value []byte) (uint64, error) {
+	if field == "data" {
+		var block rlktypes.Data
+		if err := json.Unmarshal(value, &block); err != nil {
+			return 0, err
+		}
+		return block.Height(), nil
+	} else if field == "header" {
+		var block rlktypes.SignedHeader
+		if err := json.Unmarshal(value, &block); err != nil {
+			return 0, err
+		}
+		return block.Height(), nil
+	}
+	return 0, fmt.Errorf("unknown field: %s", field)
+}
+
+type BlockFilter struct { // needs this for the Filter interface
+	start int64
+	end   int64
+	field string //need this field for differentiation between getting headers and getting data
+}
+
+func (f *BlockFilter) Filter(e dsq.Entry) bool {
+	height, err := getHeightFromEntry(f.field, e.Value)
+	if err != nil {
+		return false
+	}
+	return height >= uint64(f.end) && height <= uint64(f.start)
+}
+
+func BlockIterator(start int64, end int64) []BlockResponse {
+	var blocks []BlockResponse
+	ds, ok := env.Adapter.RollkitStore.(ds.Batching)
+	if !ok {
+		return blocks
+	}
+	filterData := &BlockFilter{start: start, end: end, field: "data"}
+	filterHeader := &BlockFilter{start: start, end: end, field: "header"}
+
+	// we need to do two queries, one for the block header and one for the block data
+	qHeader := dsq.Query{
+		Prefix: "h",
+	}
+	qHeader.Filters = append(qHeader.Filters, filterHeader)
+
+	qData := dsq.Query{
+		Prefix: "d",
+	}
+	qData.Filters = append(qData.Filters, filterData)
+	// TODO: add sorting to get the result in the right order
+
+	rHeader, err := ds.Query(context.Background(), qHeader)
+	if err != nil {
+		return blocks
+	}
+	rData, err := ds.Query(context.Background(), qData)
+	if err != nil {
+		return blocks
+	}
+
+	headerMap := make(map[uint64]*rlktypes.SignedHeader)
+	for h := range rHeader.Next() {
+		header := new(rlktypes.SignedHeader)
+		if err := header.UnmarshalBinary(h.Value); err != nil {
+			continue
+		}
+		headerMap[header.Height()] = header
+	}
+
+	dataMap := make(map[uint64]*rlktypes.Data)
+	for d := range rData.Next() {
+		data := new(rlktypes.Data)
+		if err := data.UnmarshalBinary(d.Value); err != nil {
+			continue
+		}
+		dataMap[data.Height()] = data
+	}
+
+	for height, header := range headerMap {
+		if data, ok := dataMap[height]; ok {
+			blocks = append(blocks, BlockResponse{header: header, data: data})
+		}
+	}
+
+	// Sort blocks by height descending
+	sort.Slice(blocks, func(i, j int) bool {
+		return blocks[i].header.Height() > blocks[j].header.Height()
+	})
+
+	return blocks
+}
+
+type BlockResponse struct {
+	header *rlktypes.SignedHeader
+	data   *rlktypes.Data
 }
